refactor get variables default from someone datanode
