+++ src/main/java/com/actiontech/dble/backend/mysql/nio/MySQLConnection.java
@@ -18,7 +18,7 @@ import com.actiontech.dble.route.RouteResultsetNode;
 import com.actiontech.dble.route.parser.util.Pair;
 import com.actiontech.dble.server.NonBlockingSession;
 import com.actiontech.dble.server.ServerConnection;
-import com.actiontech.dble.server.SystemVariables;
+import com.actiontech.dble.server.variables.SystemVariables;
 import com.actiontech.dble.server.parser.ServerParse;
 import com.actiontech.dble.util.StringUtil;
 import com.actiontech.dble.util.TimeUtil;
@@ -148,7 +148,7 @@ public class MySQLConnection extends BackendAIOConnection {
         this.txIsolation = -1;
         this.autocommit = true;
         //TODO:CHECK
-        this.setCharacterSet(SystemVariables.getDefaultValue("character_set_server"));
+        this.setCharacterSet(SystemVariables.getSysVars().getDefaultValue("character_set_server"));
         this.usrVariables.clear();
         this.sysVariables.clear();
     }
@@ -238,7 +238,7 @@ public class MySQLConnection extends BackendAIOConnection {
         packet.setClientFlags(clientFlags);
         packet.setMaxPacketSize(maxPacketSize);
         //TODO:CHECK
-        int charsetIndex = CharsetUtil.getCharsetDefaultIndex(SystemVariables.getDefaultValue("character_set_server"));
+        int charsetIndex = CharsetUtil.getCharsetDefaultIndex(SystemVariables.getSysVars().getDefaultValue("character_set_server"));
         packet.setCharsetIndex(charsetIndex);
 
         packet.setUser(user);
@@ -428,7 +428,7 @@ public class MySQLConnection extends BackendAIOConnection {
         }
         //tmp now = backend -(backend &&frontend)
         for (Map.Entry<String, String> entry : tmpSysVars.entrySet()) {
-            String value = SystemVariables.getDefaultValue(entry.getKey());
+            String value = SystemVariables.getSysVars().getDefaultValue(entry.getKey());
             setVars.add(new Pair<>(entry.getKey(), value));
             toResetSys.add(entry.getKey());
         }
