+++ src/main/java/com/actiontech/dble/server/handler/ServerLoadDataInfileHandler.java
@@ -20,7 +20,7 @@ import com.actiontech.dble.route.parser.druid.DruidShardingParseInfo;
 import com.actiontech.dble.route.parser.druid.RouteCalculateUnit;
 import com.actiontech.dble.route.util.RouterUtil;
 import com.actiontech.dble.server.ServerConnection;
-import com.actiontech.dble.server.SystemVariables;
+import com.actiontech.dble.server.variables.SystemVariables;
 import com.actiontech.dble.server.parser.ServerParse;
 import com.actiontech.dble.sqlengine.mpp.LoadData;
 import com.actiontech.dble.util.ObjectUtil;
@@ -116,7 +116,7 @@ public final class ServerLoadDataInfileHandler implements LoadDataInfileHandler 
         SQLTextLiteralExpr escapedExpr = (SQLTextLiteralExpr) statement.getColumnsEscaped();
         String escaped = escapedExpr == null ? "\\" : escapedExpr.getText();
         loadData.setEscape(escaped);
-        String charset = statement.getCharset() != null ? statement.getCharset() : SystemVariables.getDefaultValue("character_set_filesystem");
+        String charset = statement.getCharset() != null ? statement.getCharset() : SystemVariables.getSysVars().getDefaultValue("character_set_filesystem");
         loadData.setCharset(charset);
         loadData.setFileName(fileName);
     }
