+++ src/main/java/com/actiontech/dble/server/handler/SetHandler.java
@@ -14,7 +14,7 @@ import com.actiontech.dble.net.mysql.OkPacket;
 import com.actiontech.dble.route.parser.util.Pair;
 import com.actiontech.dble.route.parser.util.ParseUtil;
 import com.actiontech.dble.server.ServerConnection;
-import com.actiontech.dble.server.SystemVariables;
+import com.actiontech.dble.server.variables.SystemVariables;
 import com.actiontech.dble.sqlengine.OneRawSQLQueryResultHandler;
 import com.actiontech.dble.sqlengine.SetTestJob;
 import com.actiontech.dble.util.StringUtil;
@@ -231,7 +231,7 @@ public final class SetHandler {
                 if (!handleTxIsolationInMultiStmt(c, contextTask, valueExpr)) return false;
                 break;
             case SYSTEM_VARIABLES:
-                if (SystemVariables.getDefaultValue(key) == null) {
+                if (SystemVariables.getSysVars().getDefaultValue(key) == null) {
                     c.writeErrMessage(ErrorCode.ERR_NOT_SUPPORTED, "system variable " + key + " is not supported");
                 }
                 contextTask.add(new Pair<>(KeyType.SYSTEM_VARIABLES, new Pair<>(key, parseVariablesValue(valueExpr))));
@@ -361,7 +361,7 @@ public final class SetHandler {
             case TX_ISOLATION:
                 return handleTxIsolation(c, valueExpr);
             case SYSTEM_VARIABLES:
-                if (SystemVariables.getDefaultValue(key) == null) {
+                if (SystemVariables.getSysVars().getDefaultValue(key) == null) {
                     c.writeErrMessage(ErrorCode.ERR_NOT_SUPPORTED, "system variable " + key + " is not supported");
                     return false;
                 }
